# [optimize] add two stream norm for qwen3