<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
overlap RMSNorm by two stream
Co-authored-by: @ispobock

```
python3 -m sglang.launch_server  --model-path Qwen/Qwen3-235B-A22B-FP8  --tp 4
```
before
| max_concurrency | input_throughput | output_throughput | mean_ttft_ms  | median_ttft_ms | p99_ttft_ms  | mean_tpot_ms | median_tpot_ms | p99_tpot_ms | per_user_throughput |
|-----------------|------------------|-------------------|---------------|----------------|--------------|--------------|----------------|-------------|---------------------|
| 1               | 85.259           | 85.259            | 312.854       | 170.097        | 857.909      | 11.424       | 11.426         | 11.427      | 85.259              |
| 4               | 276.768          | 276.768           | 607.018       | 337.300        | 2046.613     | 13.855       | 13.765         | 15.289      | 69.192              |
| 16              | 881.044          | 881.044           | 839.877       | 750.888        | 1422.681     | 17.329       | 17.320         | 18.232      | 55.065              |
| 32              | 1417.409         | 1417.409          | 1012.238      | 1008.038       | 1340.371     | 21.574       | 21.744         | 22.633      | 44.294              |

after
| max_concurrency | input_throughput | output_throughput | mean_ttft_ms  | median_ttft_ms | p99_ttft_ms  | mean_tpot_ms | median_tpot_ms | p99_tpot_ms | per_user_throughput |
|-----------------|------------------|-------------------|---------------|----------------|--------------|--------------|----------------|-------------|---------------------|
| 1               | 86.331           | 86.331            | 318.474       | 180.309        | 852.678      | 11.273       | 11.273         | 11.276      | 86.331              |
| 4               | 282.601          | 282.601           | 470.780       | 351.867        | 1753.514     | 13.693       | 13.618         | 14.866      | 70.650              |
| 16              | 887.323          | 887.323           | 963.215       | 741.265        | 1688.938     | 17.077       | 17.139         | 18.232      | 55.458              |
| 32              | 1436.963         | 1436.963          | 1035.066      | 1132.024       | 1425.767     | 21.243       | 21.268         | 22.233      | 44.905              |

accuracy
```
python3 benchmark/gsm8k/bench_sglang.py --num-questions 1000
Accuracy: 0.948                                                                                            
Invalid: 0.001                                                                                             
Latency: 144.696 s                                                                                         
Output throughput: 965.706 token/s
```

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.